requirements.txt
uvicorn
python-fasthtml
pydantic
SQLAlchemy
fastsql
pandas
MonsterUI
pymongo